+++ svm/src/conformance/mod.rs
@@ -22,3 +22,5 @@ pub mod transaction_meta;
 pub mod txn;
 #[cfg(feature = "conformance")]
 pub mod versioned_message;
+#[cfg(feature = "conformance")]
+pub mod versioned_transaction;
